Make it to compile without warnings on x86_64.

+++ user/drbdmeta.c
@@ -1211,7 +1211,7 @@ int v07_style_md_open(struct format *cfg)
 	cfg->bd_size = bdev_size(cfg->md_fd);
 	if ((cfg->bd_size >> 9) < MD_BM_OFFSET_07) {
 		fprintf(stderr, "%s is only %llu bytes. That's not enough.\n",
-			cfg->md_device_name, cfg->bd_size);
+			cfg->md_device_name, (long long unsigned)cfg->bd_size);
 		exit(10);
 	}
 	cfg->md_offset =
@@ -1222,7 +1222,7 @@ int v07_style_md_open(struct format *cfg)
 			"byte offset %lld, but %s is only %llu bytes.\n",
 			(signed long long)cfg->md_offset,
 			cfg->md_device_name,
-			cfg->bd_size);
+			(long long unsigned)cfg->bd_size);
 		exit(10);
 	}
 
@@ -2028,9 +2028,9 @@ void check_for_existing_data(struct format *cfg)
 			     cfg->bm_offset )) >> 10;
 #undef min
 
-		printf("md_offset %llu\n", cfg->md_offset);
-		printf("al_offset %llu\n", cfg->al_offset);
-		printf("bm_offset %llu\n", cfg->bm_offset);
+		printf("md_offset %llu\n", (long long unsigned)cfg->md_offset);
+		printf("al_offset %llu\n", (long long unsigned)cfg->al_offset);
+		printf("bm_offset %llu\n", (long long unsigned)cfg->bm_offset);
 
 		/* looks like file system data */
 		printf("which uses "U64" kB\n", fs_kB);
@@ -2067,11 +2067,12 @@ void check_internal_md_flavours(struct format * cfg) {
 	flex_offset = v07_style_md_get_byte_offset(
 		DRBD_MD_INDEX_FLEX_INT, cfg->bd_size);
 
-	printf("%lld\n%lld\n%lld\n", cfg->bd_size, fixed_offset, flex_offset);
+	printf("%lld\n%lld\n%lld\n", (long long unsigned)cfg->bd_size,
+	       (long long unsigned)fixed_offset, (long long unsigned)flex_offset);
 	if (0 <= fixed_offset && fixed_offset < (off_t)cfg->bd_size - 4096) {
 		/* ... v07 fixed-size internal meta data? */
 		PREAD(cfg->md_fd, on_disk_buffer, 4096, fixed_offset);
-	
+
 		md_disk_07_to_cpu(&md_07,
 			(struct md_on_disk_07*)on_disk_buffer);
 		have_fixed_v07 = is_valid_md(Drbd_07,
@@ -2079,7 +2080,7 @@ void check_internal_md_flavours(struct format * cfg) {
 	}
 
 	PREAD(cfg->md_fd, on_disk_buffer, 4096, flex_offset);
-	
+
 	/* ... v07 (plus) flex-internal meta data? */
 	md_disk_07_to_cpu(&md_07p, (struct md_on_disk_07*)on_disk_buffer);
 	have_flex_v07 = is_valid_md(Drbd_07,
@@ -2104,17 +2105,17 @@ void check_internal_md_flavours(struct format * cfg) {
 	if (have_fixed_v07) {
 		fprintf(stderr, "There apears to be a v07 fixed-size internal meta data block\n"
 				"already in place on %s at byte offset %llu\n",
-				cfg->md_device_name, fixed_offset);
+				cfg->md_device_name, (long long unsigned)fixed_offset);
 	}
 	if (have_flex_v07) {
 		fprintf(stderr, "There apears to be a v07(plus) flexible-size internal meta data block\n"
 				"already in place on %s at byte offset %llu",
-		cfg->md_device_name, flex_offset);
+		cfg->md_device_name, (long long unsigned)flex_offset);
 	}
 	if (have_flex_v08) {
 		fprintf(stderr, "There apears to be a v08 flexible-size internal meta data block\n"
 				"already in place on %s at byte offset %llu",
-		cfg->md_device_name, flex_offset);
+		cfg->md_device_name, (long long unsigned)flex_offset);
 	}
 
 	if (have_fixed_v07 && have_flex_v07) {
